# Optimize hex encoding for UUID generation