Claim some minor memory allocation wins by using hex.Encode to generate
the final string instead of fmt.Sprintf.
```
name                      old time/op    new time/op    delta
GenerateUUID-3              2.33µs ± 1%    1.99µs ± 1%  -14.33%  (p=0.008 n=5+5)

name                      old alloc/op   new alloc/op   delta
GenerateUUID-3                184B ± 0%       64B ± 0%  -65.22%  (p=0.008 n=5+5)

name                      old allocs/op  new allocs/op  delta
GenerateUUID-3                7.00 ± 0%      2.00 ± 0%  -71.43%  (p=0.008 n=5+5)
```